Create separate test_plot.py
